+++ src/main/java/org/folio/rest/impl/TransactionAPI.java
@@ -1,6 +1,6 @@
 package org.folio.rest.impl;
 
-import static org.folio.rest.transaction.AbstractTransactionHandler.TRANSACTION_TABLE;
+import static org.folio.rest.service.AbstractTransactionService.TRANSACTION_TABLE;
 
 import java.util.Map;
 
@@ -11,10 +11,10 @@ import org.folio.rest.jaxrs.model.Transaction;
 import org.folio.rest.jaxrs.model.TransactionCollection;
 import org.folio.rest.jaxrs.resource.FinanceStorageTransactions;
 import org.folio.rest.persist.PgUtil;
-import org.folio.rest.transaction.DefaultTransactionHandler;
-import org.folio.rest.transaction.InvoiceTransactionsHandler;
-import org.folio.rest.transaction.OrderTransactionsHandler;
-import org.folio.rest.transaction.TransactionHandler;
+import org.folio.rest.service.DefaultTransactionHandler;
+import org.folio.rest.service.PaymentCreditAllOrNothingService;
+import org.folio.rest.service.OrderTransactionsService;
+import org.folio.rest.service.TransactionService;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
@@ -33,7 +33,7 @@ public class TransactionAPI implements FinanceStorageTransactions {
   @Override
   @Validate
   public void postFinanceStorageTransactions(String lang, Transaction transaction, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
-    getCreateTransactionHandler(transaction, okapiHeaders, vertxContext, asyncResultHandler).createTransaction(transaction);
+    getCreateTransactionHandler(transaction, okapiHeaders, vertxContext, asyncResultHandler).createTransaction(transaction, vertxContext, okapiHeaders);
   }
 
   @Override
@@ -53,22 +53,22 @@ public class TransactionAPI implements FinanceStorageTransactions {
   @Validate
   public void putFinanceStorageTransactionsById(String id, String lang, Transaction transaction, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
     transaction.setId(id);
-    getUpdateTransactionHandler(transaction, okapiHeaders, vertxContext, asyncResultHandler).updateTransaction(transaction);
+    getUpdateTransactionHandler(transaction, okapiHeaders, vertxContext, asyncResultHandler).updateTransaction(transaction, vertxContext, okapiHeaders);
   }
 
-  private TransactionHandler getCreateTransactionHandler(Transaction transaction, Map<String, String> okapiHeaders, Context vertxContext, Handler<AsyncResult<Response>> asyncResultHandler) {
+  private TransactionService getCreateTransactionHandler(Transaction transaction, Map<String, String> okapiHeaders, Context vertxContext, Handler<AsyncResult<Response>> asyncResultHandler) {
     if (transaction.getTransactionType() == Transaction.TransactionType.ENCUMBRANCE) {
-      return new OrderTransactionsHandler(okapiHeaders, vertxContext, asyncResultHandler);
+      return new OrderTransactionsService(okapiHeaders, vertxContext, asyncResultHandler);
     } else if (transaction.getTransactionType() == Transaction.TransactionType.PAYMENT
         || transaction.getTransactionType() == Transaction.TransactionType.CREDIT) {
-      return new InvoiceTransactionsHandler(okapiHeaders, vertxContext, asyncResultHandler);
+      return new PaymentCreditAllOrNothingService(okapiHeaders, vertxContext, asyncResultHandler);
     }
     return new DefaultTransactionHandler(okapiHeaders, vertxContext, asyncResultHandler);
   }
 
-  private TransactionHandler getUpdateTransactionHandler(Transaction transaction, Map<String, String> okapiHeaders, Context vertxContext, Handler<AsyncResult<Response>> asyncResultHandler) {
+  private TransactionService getUpdateTransactionHandler(Transaction transaction, Map<String, String> okapiHeaders, Context vertxContext, Handler<AsyncResult<Response>> asyncResultHandler) {
     if (transaction.getTransactionType() == Transaction.TransactionType.ENCUMBRANCE) {
-      return new InvoiceTransactionsHandler(okapiHeaders, vertxContext, asyncResultHandler);
+      return new PaymentCreditAllOrNothingService(okapiHeaders, vertxContext, asyncResultHandler);
     }
     return new DefaultTransactionHandler(okapiHeaders, vertxContext, asyncResultHandler);
   }
