+++ src/main/java/org/folio/rest/persist/Tx.java
@@ -3,6 +3,7 @@ package org.folio.rest.persist;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
+import io.vertx.core.Vertx;
 import io.vertx.ext.sql.SQLConnection;
 
 public class Tx<T> {
@@ -10,6 +11,7 @@ public class Tx<T> {
   private T entity;
   private PostgresClient pgClient;
   private AsyncResult<SQLConnection> sqlConnection;
+
 
   public Tx(T entity, PostgresClient pgClient) {
     this.entity = entity;
