+++ src/main/java/org/folio/rest/service/BudgetService.java
@@ -1,15 +1,21 @@
 package org.folio.rest.service;
 
 import static javax.ws.rs.core.Response.Status.NOT_FOUND;
+import static org.folio.rest.impl.LedgerAPI.LEDGER_TABLE;
 import static org.folio.rest.persist.HelperUtils.getFullTableName;
 import static org.folio.rest.util.ResponseUtils.handleFailure;
 import static org.folio.rest.util.ResponseUtils.handleNoContentResponse;
 
 import javax.ws.rs.core.Response;
 
+import org.folio.rest.dao.BudgetDAO;
 import org.folio.rest.jaxrs.model.Budget;
+import org.folio.rest.jaxrs.model.Ledger;
+import org.folio.rest.jaxrs.model.LedgerCollection;
 import org.folio.rest.jaxrs.model.Transaction;
+import org.folio.rest.jaxrs.resource.FinanceStorageLedgers;
 import org.folio.rest.persist.CriterionBuilder;
+import org.folio.rest.persist.PgUtil;
 import org.folio.rest.persist.PostgresClient;
 import org.folio.rest.persist.Tx;
 import org.folio.rest.persist.Criteria.Criterion;
@@ -26,6 +32,10 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.web.handler.impl.HttpStatusException;
+import org.folio.rest.tools.utils.TenantTool;
+import org.javamoney.moneta.Money;
+
+import java.util.Map;
 
 public class BudgetService {
 
@@ -33,10 +43,13 @@ public class BudgetService {
   private static final String GROUP_FUND_FY_TABLE = "group_fund_fiscal_year";
   private static final String TRANSACTIONS_TABLE = "transaction";
   public static final String TRANSACTION_IS_PRESENT_BUDGET_DELETE_ERROR = "transactionIsPresentBudgetDeleteError";
+  public static final String BUDGET_IS_INACTIVE = "Cannot create encumbrance from the not active budget";
+  public static final String FUND_CANNOT_BE_PAID = "Fund cannot be paid due to restrictions";
 
   private final Logger logger = LoggerFactory.getLogger(this.getClass());
   private final String tenantId;
   private PostgresClient pgClient;
+  private BudgetDAO budgetDAO;
 
   public BudgetService(Vertx vertx, String tenantId) {
     this.tenantId = tenantId;
@@ -59,6 +72,130 @@ public class BudgetService {
             }))
           .onComplete(handleNoContentResponse(asyncResultHandler, id, "Budget {} {} deleted"));
       });
+  }
+
+  public Future<Void> verifyBudgetHasEnoughMoney(Transaction transaction, Context context, Map<String, String> headers) {
+    String fundId = transaction.getTransactionType() == Transaction.TransactionType.CREDIT ? transaction.getToFundId() : transaction.getFromFundId();
+    PostgresClient client = PostgresClient.getInstance(context.owner(), TenantTool.tenantId(headers));
+    return budgetDAO.getBudgetByFundIdAndFiscalYearId(transaction.getFiscalYearId(), fundId, client)
+      .compose(budget -> checkTransactionRestrictions(transaction, budget));
+  }
+
+
+  private Future<Void> checkTransactionRestrictions(Transaction transaction, Budget budget) {
+    Promise<Void> promise = Promise.promise();
+
+    if (budget.getBudgetStatus() != Budget.BudgetStatus.ACTIVE) {
+      logger.error(BUDGET_IS_INACTIVE, budget.getId());
+      promise.fail(new HttpStatusException(Response.Status.BAD_REQUEST.getStatusCode(), BUDGET_IS_INACTIVE));
+    } else {
+      return checkLedgerRestrictions(transaction, budget);
+    }
+    return promise.future();
+  }
+
+  private Future<Void> checkLedgerRestrictions(Transaction transaction, Budget budget) {
+    Promise<Void> promise = Promise.promise();
+    getExistentLedger(transaction).onComplete(result -> {
+      if (result.succeeded()) {
+        if (isEncumbranceRestricted(transaction, budget, result.result())) {
+          Money remainingAmountForEncumbrance = getBudgetRemainingAmountForEncumbrance(budget, transaction.getCurrency());
+          if (Money.of(transaction.getAmount(), transaction.getCurrency()).isGreaterThan(remainingAmountForEncumbrance)) {
+            promise.fail(new HttpStatusException(Response.Status.BAD_REQUEST.getStatusCode(), FUND_CANNOT_BE_PAID));
+            return;
+          }
+        }
+        else if (isPaymentRestricted(transaction, budget, result.result())) {
+          Money remainingAmountForPayment = getBudgetRemainingAmountForPayment(budget, transaction.getCurrency());
+          if (Money.of(transaction.getAmount(), transaction.getCurrency()).isGreaterThan(remainingAmountForPayment)) {
+            promise.fail(new HttpStatusException(Response.Status.BAD_REQUEST.getStatusCode(), FUND_CANNOT_BE_PAID));
+            return;
+          }
+        }
+        promise.complete();
+      } else {
+        handleFailure(promise, result);
+      }
+    });
+    return promise.future();
+  }
+
+
+  private boolean isEncumbranceRestricted(Transaction transaction, Budget budget, Ledger ledger) {
+    return transaction.getTransactionType() == Transaction.TransactionType.ENCUMBRANCE && ledger.getRestrictEncumbrance()
+      && budget.getAllowableEncumbrance() != null;
+  }
+
+  /**
+   * Calculates remaining amount for encumbrance
+   * [remaining amount] = (allocated * allowableEncumbered) - (allocated - (unavailable + available)) - (encumbered + awaitingPayment + expenditures)
+   *
+   * @param budget     processed budget
+   * @param currency   processed transaction currency
+   * @return remaining amount for encumbrance
+   */
+  private Money getBudgetRemainingAmountForEncumbrance(Budget budget, String currency) {
+    Money allocated = Money.of(budget.getAllocated(), currency);
+    // get allowableEncumbered converted from percentage value
+    double allowableEncumbered = Money.of(budget.getAllowableEncumbrance(), currency).divide(100d).getNumber().doubleValue();
+    Money unavailable = Money.of(budget.getUnavailable(), currency);
+    Money available = Money.of(budget.getAvailable(), currency);
+    Money encumbered = Money.of(budget.getEncumbered(), currency);
+    Money awaitingPayment = Money.of(budget.getAwaitingPayment(), currency);
+    Money expenditures = Money.of(budget.getExpenditures(), currency);
+
+    Money result = allocated.multiply(allowableEncumbered);
+    result = result.subtract(allocated.subtract(unavailable.add(available)));
+    result = result.subtract(encumbered.add(awaitingPayment).add(expenditures));
+
+    return result;
+  }
+
+  /**
+   * Calculates remaining amount for payment
+   * [remaining amount] = (allocated * allowableExpenditure) - (allocated - (unavailable + available)) - (awaitingPayment + expended)
+   *
+   * @param budget     processed budget
+   * @param currency   processed transaction currency
+   * @return remaining amount for payment
+   */
+  private Money getBudgetRemainingAmountForPayment(Budget budget, String currency) {
+    Money allocated = Money.of(budget.getAllocated(), currency);
+    // get allowableExpenditure from percentage value
+    double allowableExpenditure = Money.of(budget.getAllowableExpenditure(), currency).divide(100d).getNumber().doubleValue();
+    Money unavailable = Money.of(budget.getUnavailable(), currency);
+    Money available = Money.of(budget.getAvailable(), currency);
+    Money expenditure = Money.of(budget.getExpenditures(), currency);
+    Money awaitingPayment = Money.of(budget.getAwaitingPayment(), currency);
+
+    Money result = allocated.multiply(allowableExpenditure);
+    result = result.subtract(allocated.subtract(unavailable.add(available)));
+    result = result.subtract(expenditure.add(awaitingPayment));
+
+    return result;
+  }
+
+  private boolean isPaymentRestricted(Transaction transaction, Budget budget, Ledger ledger) {
+    return transaction.getTransactionType() == Transaction.TransactionType.PAYMENT && ledger.getRestrictExpenditures()
+      && budget.getAllowableExpenditure() != null;
+  }
+
+  private Future<Ledger> getExistentLedger(Transaction transaction) {
+    Promise<Ledger> promise = Promise.promise();
+    String fundId = transaction.getTransactionType() == Transaction.TransactionType.CREDIT ? transaction.getToFundId() : transaction.getFromFundId();
+    String query = "fund.id =" + fundId;
+    PgUtil.get(LEDGER_TABLE, Ledger.class, LedgerCollection.class, query , 0, 1, getOkapiHeaders(), getVertxContext(),
+      FinanceStorageLedgers.GetFinanceStorageLedgersResponse.class, reply-> {
+        if (reply.failed()) {
+          handleFailure(promise, reply);
+        } else if (((LedgerCollection) reply.result().getEntity()).getLedgers().isEmpty()) {
+          logger.error(LEDGER_NOT_FOUND_FOR_TRANSACTION);
+          promise.fail(new HttpStatusException(Response.Status.BAD_REQUEST.getStatusCode(), LEDGER_NOT_FOUND_FOR_TRANSACTION));
+        } else {
+          promise.complete(((LedgerCollection) reply.result().getEntity()).getLedgers().get(0));
+        }
+      });
+    return promise.future();
   }
 
   private Future<Tx<String>> deleteBudget(Tx<String> tx) {
@@ -137,4 +274,6 @@ public class BudgetService {
     });
     return promise.future();
   }
+
+
 }
