+++ src/test/java/org/folio/rest/impl/EncumbrancesTest.java
@@ -6,10 +6,10 @@ import static org.folio.rest.impl.TransactionTest.LEDGER_FYS_ENDPOINT;
 import static org.folio.rest.impl.TransactionTest.TRANSACTION_TENANT_HEADER;
 import static org.folio.rest.impl.TransactionsSummariesTest.INVOICE_TRANSACTION_SUMMARIES_ENDPOINT;
 import static org.folio.rest.impl.TransactionsSummariesTest.ORDER_TRANSACTION_SUMMARIES_ENDPOINT;
-import static org.folio.rest.transaction.AllOrNothingHandler.ALL_EXPECTED_TRANSACTIONS_ALREADY_PROCESSED;
-import static org.folio.rest.transaction.AllOrNothingHandler.BUDGET_IS_INACTIVE;
-import static org.folio.rest.transaction.AllOrNothingHandler.BUDGET_NOT_FOUND_FOR_TRANSACTION;
-import static org.folio.rest.transaction.AllOrNothingHandler.FUND_CANNOT_BE_PAID;
+import static org.folio.rest.service.AllOrNothingTransactionService.ALL_EXPECTED_TRANSACTIONS_ALREADY_PROCESSED;
+import static org.folio.rest.service.AllOrNothingTransactionService.BUDGET_IS_INACTIVE;
+import static org.folio.rest.dao.BudgetDAO.BUDGET_NOT_FOUND_FOR_TRANSACTION;
+import static org.folio.rest.service.AllOrNothingTransactionService.FUND_CANNOT_BE_PAID;
 import static org.folio.rest.utils.TenantApiTestUtil.deleteTenant;
 import static org.folio.rest.utils.TenantApiTestUtil.prepareTenant;
 import static org.folio.rest.utils.TestEntities.BUDGET;
